Move the region_meta module to top level
