--- docs/source/api/atlannot.atlas.region_meta.rst
@@ -1,7 +1,0 @@
-atlannot.atlas.region\_meta module
-==================================
-
-.. automodule:: atlannot.atlas.region_meta
-   :members:
-   :undoc-members:
-   :show-inheritance:
