+++ docs/source/api/atlannot.region_meta.rst
@@ -1,0 +1,7 @@
+atlannot.region\_meta module
+============================
+
+.. automodule:: atlannot.region_meta
+   :members:
+   :undoc-members:
+   :show-inheritance:
