+++ docs/source/api/atlannot.rst
@@ -19,6 +19,7 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
+   atlannot.region_meta
    atlannot.utils
 
 Module contents
