+++ src/atlannot/atlas/__init__.py
@@ -13,10 +13,8 @@
 # limitations under the License.
 """Module for working with annotation atlases."""
 from .align import get_misalignment, unfurl_regions, warp
-from .region_meta import RegionMeta
 
 __all__ = [
-    "RegionMeta",
     "get_misalignment",
     "unfurl_regions",
     "warp",
