+++ src/atlannot/merge/fine.py
@@ -26,7 +26,7 @@ from collections import deque
 import numpy as np
 from numpy import ma
 
-from atlannot.atlas.region_meta import RegionMeta
+from atlannot.region_meta import RegionMeta
 from atlannot.merge.common import atlas_remap, descendants, replace
 
 logger = logging.getLogger(__name__)
