+++ tests/test_atlas_align.py
@@ -19,7 +19,7 @@ import pytest
 from tqdm import tqdm
 
 from atlannot.atlas.align import get_misalignment, specific_label_iou, unfurl_regions
-from atlannot.atlas.region_meta import RegionMeta
+from atlannot.region_meta import RegionMeta
 
 
 class TestAlign:
